feat: 일기 관련 Mock Controller 코드리뷰 피드백 적용 (#2)

+++ src/main/java/com/dnd/bbok/domain/diary/dto/response/DiaryChecklistDto.java
@@ -6,12 +6,12 @@ import java.util.Random;
 
 @Getter
 public class DiaryChecklistDto {
-    private final int id;
+    private final Long id;
     private final String criteria;
     private final boolean isChecked;
 
     public DiaryChecklistDto(String criteria, boolean isChecked) {
-        this.id = new Random().nextInt(21);
+        this.id = new Random().nextLong(21);
         this.criteria = criteria;
         this.isChecked = isChecked;
     }
